Moves CompactFeature component to ui folder
