+++ components/ui/CompactFeature.vue
